yolo_model.py
import torch
import numpy as np
import cv2
import io
from PIL import Image
from ultralytics import YOLO
from ultralytics.models import YOLOv10



test_image = cv2.imread('test.png')

model_object_detect = YOLO('models/yolov8x-oiv7.pt')
model_face_emmotion = YOLOv10('models/yolov10n-face.pt')

def detect_objects(image: Image.Image):
    # PIL.Image를 numpy 배열로 변환
    np_image = np.array(image)

    # 모델을 사용하여 감지
    results_object = model_object_detect(np_image)
    results_face_emotion = model_face_emmotion(np_image)

    # 결과를 이미지로 변환
    od_image = results_object[0].plot()
    fc_image = results_face_emotion[0].plot()

    # numpy 배열을 PIL.Image로 변환
    od_image_pil = Image.fromarray(od_image)
    fc_image_pil = Image.fromarray(fc_image)

    # 이미지를 바이트 형태로 변환
    output1 = io.BytesIO()
    output2 = io.BytesIO()

    od_image_pil.save(output1, format="PNG")
    fc_image_pil.save(output2, format="PNG")

    return output1.getvalue(), output2.getvalue()

# 테스트
if __name__ == "__main__":
    # PIL.Image로 변환
    image = Image.fromarray(test_image)

    # 감지 함수 호출
    output1, output2 = detect_objects(image)

    # 결과를 파일로 저장 (테스트용)
    with open('test/od_output.png', 'wb') as f:
        f.write(output1)
    
    with open('test/fc_output.png', 'wb') as f:
        f.write(output2)